small fixes
The which-key mapping for the lazygit in Vim wasn't set properly so I fix it.
I removed the pyright configuration from the coc-settings.json because of a poor user experience with the signs. I added the pylsp in the place.
I added the NvChad to the fzf function for selecting the nvim setup. I'll try to configure it later to make let it available in other repository.